+++ scplint/__init__.py
@@ -3,5 +3,5 @@
 '''
 
 __author__ = 'Richard Zimmermann'
-__version__ = '0.0.2'
+__version__ = '0.0.3'
 __all__ = ['aws_actions', 'log', 'report', 'scp', 'statement']